refactor(core): share the built-in video model catalog
The video model catalog lived inside the API generation service, so the web
composer had no way to know which aspect ratios, durations, or resolutions a
model accepts. Move it to @vm0/core unchanged and import it back.

Also names the three request defaults the service applied as inline literals
(model, aspect ratio, duration) so both sides resolve them from one place.

Pure refactor: every moved line is byte-identical and the service keeps its
provider request shapes, pricing categories, and output dimension tables.
